==> 04_report_agencyteams.Rmd <==
---
title: "Biden Transition - Agency Review Teams"
author: "Aaron Kessler, Data/Investigations Team"
date: "`r format(Sys.time(), '%B %d, %Y')`"
output: 
  html_document:
    toc: true
    toc_float: true
    code_folding: hide
    # code_download: true
    # includes:
    #   after_body: footer.html
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(lubridate)
library(janitor)
library(scales)
library(gt)
library(kableExtra)
options(dplyr.summarise.inform = FALSE)

#run the script for step 03 to update data files
source("03_analysis.R")


#load data files from step 03
#entire combined agency teams file
agencyteams <- readRDS("processed_data/agencyteams.rds")
#names of new agency review team members
newnames <- readRDS("processed_data/newnames.rds")
#aggregate county of agency totals compared
agencycount_compare <- readRDS("processed_data/agencycount_compare.rds")


```

####
The following is an analysis based on scraping Joe Biden's presidential transition [website](https://buildbackbetter.gov/the-transition/agency-review-teams/) to capture the various lists of members on the Agency Review Teams.  
  
The names below reflect the currently available data as of **`r format(Sys.time(), '%I:%M %p on %b. %d')`**.


####
### *New Names*

Here are the new names added to the agency review team lists since last time.  
Note that some people may already be part of another team and/or assigned to multiple agency teams - they're marked below, as applicable.  
  

```{r, echo=FALSE}

newnames %>% 
  select(agency, name, most_recent_employment, on_multiple_teams, team_lead) %>% 
  gt() %>%
  tab_options(table.align = "left") %>%
  cols_align(
    align = "center",
    columns = vars(team_lead, on_multiple_teams)
  ) 
 

```


####
### *Aggregate Team Changes*

The total number of people appointed to each agency team, along with change since last time reflecting the number of new people added. (Agencies with no change are omitted here.)
  

```{r, echo=FALSE}

agencycount_compare %>% 
  filter(change > 0) %>% 
  mutate(
    change = paste0("+", change)
  ) %>% 
  gt() %>%
  tab_options(table.align = "left") %>% 
 tab_style(
    style = list(
      # cell_fill(color = "lightcyan"),
      cell_text(weight = "bold",
                color = "darkgreen")
      ),
    locations = cells_body(
      columns = vars(change)
      )
  ) %>% 
  tab_style(
    style = list(
      cell_text(color = "gray")
      ),
    locations = cells_body(
      columns = vars(previous)
      )
  ) 
  
 

```


####
### *Largest and Smallest Overall*

The **largest** agency review teams as of today are:

```{r, echo=FALSE}

agencycount_compare %>% 
  select(agency, current) %>% 
  arrange(desc(current)) %>% 
  head(8) %>% 
  gt() %>%
  tab_options(table.align = "left") 
 

```

####
The **smallest** agency review teams (less than six members) as of today are:

```{r, echo=FALSE}

agencycount_compare %>% 
  select(agency, current) %>% 
  arrange(current) %>% 
  filter(current < 6) %>% 
  gt() %>%
  tab_options(table.align = "left") 
 

```


####
### *Top Companies/Organizations Represented*

Here are the most frequent recent employers of agency review team members. 
  
Because some members are on multiple teams, we've accounted for that to count each person only once for employer totals. (E.g. if John Doe works for Google and is on three separate teams, Google gets one person added to its total, not three.)  

```{r, echo=FALSE}

#because some people are appointed to one team, we'll want to account
#for that when looking at top employers, to avoid double/triple counting

agencyteams %>% 
  select(namestring, most_recent_employment_standardized) %>% 
  distinct() %>% 
  count(most_recent_employment_standardized, sort = TRUE) %>% 
  filter(n >= 3) %>% 
  gt() %>%
  tab_options(table.align = "left") 
 

```

####
**California people**

```{r, echo=FALSE}

agencyteams %>% 
  filter(str_detect(most_recent_employment_standardized, "CALIFORNIA")) %>% 
  arrange(name) %>% 
  write_csv("output/california.csv")
  
 

```
